requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
dateparser==0.7.2
Django==3.0.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7

MarkupSafe-1.1.1 asgiref-3.2.3 certifi-2019.11.28 chardet-3.0.4 coreapi-2.3.3 coreschema-0.0.4 django-3.0.3 django-rest-swagger-2.2.0 djangorestframework-3.11.0 idna-2.8 itypes-1.1.0 jinja2-2.11.1 openapi-codec-1.3.2 pytz-2019.3 requests-2.22.0 simplejson-3.17.0 sqlparse-0.3.0 uritemplate-3.0.1 urllib3-1.25.8